Final changes

+++ src/main/java/beitech/test/order/application/ordersApplication/Persistence/OrdersRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by alejandra on 15/10/17.
  */
-public interface OrdersRepository extends JpaRepository<Order, Integer>{
+public interface OrdersRepository extends JpaRepository<Order, Long>{
 }
